Move wiki and web stories components into learning folder, create dummy helpcenter and learning components, create sidebar data object with learning and it's children.
